Refactor: move operations under standalone module
